fix: Prevent general IP forwarding
Ella Core enables IP forwarding in the Linux kernel to enable the
XDP implementation of the UPF to work. While necessary, this
creates a potential security issue where the kernel will happily
forward between any network interfaces. This would make it possible for
an attacker with network access to any interface to gain access to
other attached networks.

This fix uses nftables to add a default policy to drop all forwarding.
Because the UPF is implemented with XDP, this bypasses the firewall
and the functionality is maintained.

An alternative would be to document the need to setup the firewall rules
before running Ella Core, but this would be risky. The fact that we
enable IP forwarding ourselves makes it our responsibility to prevent
this issue in my view.