feat: Add VLAN support in XDP native mode
This changeset introduces support for VLANs in XDP native mode. It was
already possible to use vlan interfaces in generic mode, but the core
would fail to start when trying to use native mode, as the vlan driver
does not support XDP.

The approach taken here is to inspect the provided configuration,
and when a combination of native mode and vlan interfaces is detected,
we attach the XDP program to the parent interface instead. We also
identify the VLAN ID from the vlan interface and pass it as
a configuration to the XDP program so that it can insert it in outgoing
packets.

We also provide the interface index of the N3 and N6 attachement
interfaces (parent interface in vlan + native, vlan interface in vlan
+ generic) to the XDP program. This lets the program route the packet
through the correct interface, overriding the interface received by
the bpf_fib_lookup function, as it would return the index of the vlan
interface in native mode, causing Operation Unsupported errors.

This has been tested on both N3 and N6, in both generic and native mode,
with VLAN hardware offload enabled or disabled on the interfaces.

This is currently incomplete and is missing the following elements:

- Support N3 and N6 as VLANs on the same parent interface