fix: Properly send Configuration Update Command
Ella Core previously did not send a Configuration Update Command
to the UE for changing the GUTI after a successful network initiated
Service Request procedure.

This changeset fixes the issue by adding support for sending this
message and handling its response. The current implementation is very
basic and was only tested for this particular use case.

We also added a way to keep track of an old GUTI, as both are the old
and new GUTI should be considered valid by the AMF until acknowledged
by the UE with a ConfigurationUpdateComplete.